Move AboutPagerAdapter outside AboutActivity
